auth/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
flask-marshmallow==0.14.0
pydantic==1.9.2
marshmallow-sqlalchemy==0.28.1
bcrypt==3.2.2
cryptography==37.0.4
PyJWT==2.4.0
gevent==21.12.0
redis==4.3.4
apispec==5.2.2
apispec_webframeworks==0.5.2
flask-swagger-ui==4.11.1
requests==2.27.1
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
pyyaml==6.0
ua-parser==0.16.1
user-agents==2.2.0
flask-limiter==2.6.2
marshmallow_enum==1.5.1
grpcio==1.53.0
grpcio-tools==1.53.0 
